requirements.txt
streamlit==1.21.0
chromadb==0.1.11
langchain==0.17.3
langchain-groq==0.1.1
huggingface-hub==0.15.1
huggingface-transformers==4.32.0
python-dotenv==1.0.0
uuid==1.30
re==2.2.1
